+++ fj-doc-mod-fop/src/main/java/org/fugerit/java/doc/mod/fop/PdfFopTypeNoAccessibilityHandler.java
@@ -1,0 +1,9 @@
+package org.fugerit.java.doc.mod.fop;
+
+public class PdfFopTypeNoAccessibilityHandler extends PdfFopTypeHandler {
+
+    public PdfFopTypeNoAccessibilityHandler() {
+        super(  Boolean.FALSE, DEFAULT_KEEP_EMPTY_TAGS );
+    }
+
+}